Quitar todas las funciones AJAX - HECHO Borrar jugadores al acabar partida - HECHO Que se pueda jugar sin actualizar servidor - HECHO Reiniciar mapa al iniciar otra partida - HECHO Texto de matching no aparece si echas otra partida - SOLVED No se borran los sprites al recoger los objetos (¿solo ocurre a veces? ¿WTF?) - SIN RESOLVER No bajan la vida de los jugadores al pegar puñetazos - SIN RESOLVER

+++ DLabyrinth_WS/src/main/java/es/urjc/ramstudios/dlabyrinth/Bala.java
@@ -20,15 +20,15 @@ public class Bala {
 		return directionX;
 	}
 
-	public void setDirectionX(float direccionX) {
+	public void setDirectionX(int direccionX) {
 		this.directionX = direccionX;
 	}
 
-	public int getDirectionY() {
+	public float getDirectionY() {
 		return directionY;
 	}
 
-	public void setDirectionY(float direccionY) {
+	public void setDirectionY(int direccionY) {
 		this.directionY = direccionY;
 	}
 /*
